NLP
An introduction to NLP to classify a tweet whether it belongs to hate speech or not(1&0). The tweets needs to cleaned and then features need to be generated from the tweet. Regular expression module is used to tokenize the text and set of custom defined functions are used to generate new features. The modeling part over here is pretty easy Naive Bayes is used in combination Accuracy_score as metric for model evaluation.